testing/web-platform/tests/eventsource/eventsource-request-cancellation.htm

<!DOCTYPE html>
<html>
  <head>
    <title>EventSource: request cancellation</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script>
      t = async_test();
      onload = t.step_func(function() {
        var url = "resources/message.py?sleep=1000&message=" + encodeURIComponent("retry:1000\ndata:abc\n\n");
        var es = new EventSource(url);
        var got_error = false;
        es.onerror = t.step_func(function() {
          got_error = true;
          assert_equals(es.readyState, EventSource.CLOSED)
        });

        setTimeout(function() {
            window.stop()
            es.onopen = t.unreached_func("Got open event");
            es.onmessage = t.unreached_func("Got message after closing source");
        }, 0);

        setTimeout(t.step_func(function () {
          assert_equals(es.readyState, EventSource.CLOSED,
                        "After stopping the eventsource readyState should be CLOSED")
          assert_true(got_error, "Error event should have fired");
          t.done();
        }), 4000);
      });
    </script>
  </body>
</html>
